engine: move LiveUpdateBuildAndDeployer into buildcontrol
