+++ internal/engine/buildcontrol/testdata_test.go
@@ -6,11 +6,20 @@ import (
 	"github.com/tilt-dev/tilt/internal/container"
 	"github.com/tilt-dev/tilt/internal/k8s"
 	"github.com/tilt-dev/tilt/internal/k8s/testyaml"
+	"github.com/tilt-dev/tilt/internal/store"
 	"github.com/tilt-dev/tilt/internal/testutils/manifestbuilder"
 	"github.com/tilt-dev/tilt/pkg/model"
 )
 
 type Fixture = manifestbuilder.Fixture
+
+var testImageRef = container.MustParseNamedTagged("gcr.io/some-project-162817/sancho:deadbeef")
+var imageTargetID = model.TargetID{
+	Type: model.TargetTypeImage,
+	Name: "gcr.io/some-project-162817/sancho",
+}
+
+var alreadyBuilt = store.NewImageBuildResultSingleRef(imageTargetID, testImageRef)
 
 const SanchoYAML = testyaml.SanchoYAML
