--- internal/engine/extractors.go
@@ -1,101 +1,0 @@
-package engine
-
-import (
-	"fmt"
-
-	"github.com/pkg/errors"
-
-	"github.com/tilt-dev/tilt/internal/container"
-	"github.com/tilt-dev/tilt/internal/engine/buildcontrol"
-	"github.com/tilt-dev/tilt/internal/sliceutils"
-	"github.com/tilt-dev/tilt/internal/store"
-	"github.com/tilt-dev/tilt/pkg/model"
-)
-
-// If there are images that can be updated in-place in a container, return
-// a state tree of what needs to be updated.
-func extractImageTargetsForLiveUpdates(specs []model.TargetSpec, stateSet store.BuildStateSet) ([]liveUpdateStateTree, error) {
-	g, err := model.NewTargetGraph(specs)
-	if err != nil {
-		return nil, errors.Wrap(err, "extractImageTargetsForLiveUpdates")
-	}
-
-	if !g.IsSingleSourceDAG() {
-		return nil, fmt.Errorf("Cannot extract live updates on this build graph structure")
-	}
-
-	result := make([]liveUpdateStateTree, 0)
-
-	deployedImages := g.DeployedImages()
-	for _, iTarget := range deployedImages {
-		state := stateSet[iTarget.ID()]
-		if state.IsEmpty() {
-			return nil, buildcontrol.SilentRedirectToNextBuilderf("In-place build does not support initial deploy")
-		}
-
-		if state.FullBuildTriggered {
-			return nil, buildcontrol.SilentRedirectToNextBuilderf("Force update (triggered manually, not automatically, with no dirty files)")
-		}
-
-		if len(state.DepsChangedSet) > 0 {
-			return nil, buildcontrol.SilentRedirectToNextBuilderf("Pending dependencies")
-		}
-
-		hasFileChangesIDs, err := buildcontrol.HasFileChangesTree(g, iTarget, stateSet)
-		if err != nil {
-			return nil, errors.Wrap(err, "extractImageTargetsForLiveUpdates")
-		}
-
-		// If this image and none of its dependencies need a rebuild,
-		// we can skip it.
-		if len(hasFileChangesIDs) == 0 {
-			continue
-		}
-
-		luInfo := iTarget.LiveUpdateInfo()
-		if luInfo.Empty() {
-			return nil, buildcontrol.SilentRedirectToNextBuilderf("LiveUpdate requires that LiveUpdate details be specified")
-		}
-
-		if state.RunningContainerError != nil {
-			return nil, buildcontrol.RedirectToNextBuilderInfof("Error retrieving container info: %v", state.RunningContainerError)
-		}
-
-		// Now that we have live update information, we know this CAN be updated in
-		// a container(s). Check to see if we have enough information about the
-		// container(s) that would need to be updated.
-		if len(state.RunningContainers) == 0 {
-			return nil, buildcontrol.RedirectToNextBuilderInfof("Don't have info for running container of image %q "+
-				"(often a result of the deployment not yet being ready)", container.FamiliarString(iTarget.Refs.ClusterRef()))
-		}
-
-		filesChanged, err := filesChangedTree(g, iTarget, stateSet)
-		if err != nil {
-			return nil, errors.Wrap(err, "extractImageTargetsForLiveUpdates")
-		}
-
-		result = append(result, liveUpdateStateTree{
-			iTarget:           iTarget,
-			filesChanged:      filesChanged,
-			iTargetState:      state,
-			hasFileChangesIDs: hasFileChangesIDs,
-		})
-	}
-
-	return result, nil
-}
-
-// Given a target, return all the files in its tree of dependencies that have
-// changed.
-func filesChangedTree(g model.TargetGraph, target model.TargetSpec, stateSet store.BuildStateSet) ([]string, error) {
-	result := []string{}
-	err := g.VisitTree(target, func(current model.TargetSpec) error {
-		state := stateSet[current.ID()]
-		result = append(result, state.FilesChanged()...)
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-	return sliceutils.DedupedAndSorted(result), nil
-}
